Feature/triggerreload

+++ src/model/views/manage.py
@@ -2,12 +2,39 @@ from django.shortcuts import HttpResponse, redirect
 from shutil import copyfile
 import os
 import threading
+import subprocess
 import sys
+import signal
 from ..models import *
+from ..generators import *
 from django.conf import settings
 from django.utils import autoreload
+from django.core import management
+import time
+
+def restart(request):
+    for thread in threading.enumerate():
+        os.kill(
+            thread.native_id,
+            signal.SIGHUP
+        )
+    return HttpResponse(json.dumps('Restarting Gunicorn...'))
 
 def clear(request):
+    for c in Classifier.objects.all():
+        ClassifierGenerator(c).delete()
+        c.delete()
+    for p in Property.objects.all():
+        PropertyGenerator(p).delete()
+        p.delete()
+    for o in Operation.objects.all():
+        o.delete()
+    for r in Relationship.objects.all():
+        RelationshipGenerator(r).delete()
+        r.delete()
+    for a in Application.objects.all():
+        ApplicationGenerator(a).delete()
+        a.delete()
     models = settings.BASE_DIR + '/shared/models.py'
     default = settings.BASE_DIR + '/shared/models.default.py'
     copyfile(
@@ -21,34 +48,27 @@ def clear(request):
         'otherclassifier.py'
     ]
     for f in os.listdir(settings.BASE_DIR + '/shared/views'):
-        if f not in notDelete:
-             os.remove(
-                 settings.BASE_DIR + '/shared/views/' + f
-             )
-    for c in Class.objects.all():
-        c.delete()
-    for c in Classifier.objects.all():
-        c.delete()
-    for p in Property.objects.all():
-        p.delete()
-    for o in Operation.objects.all():
-        o.delete()
-    for g in Generalization.objects.all():
-        g.delete()
-    for op in OperationParameter.objects.all():
-        op.delete()
-    for a in Association.objects.all():
-        a.delete()
-    for c in Composition.objects.all():
-        c.delete()
-    for r in Relationship.objects.all():
-        r.delete()
-    ar = autoreload.restart_with_reloader()
-    sys.exit(ar)
-    # -> THIS RESTARTS THE SERVER BUT THE PORT WILL BE IN USE
-    # -> POSSIBLE FIX?
-    return redirect('/model')
-    
+        if f not in notDelete and f:
+            if not os.path.isdir(f):
+                os.remove(
+                    settings.BASE_DIR + '/shared/views/' + f
+                )
+    restart(True)
+    return HttpResponse(json.dumps('Restarting Server...'))
+
 def generate(request):
-    # code to generate
-    return redirect('/model')
+    for c in Classifier.objects.all():
+        ClassifierGenerator(c).generate(True)
+    for r in Relationship.objects.all():
+        RelationshipGenerator(r).generate(True)
+    for p in Property.objects.all():
+        PropertyGenerator(p).generate(True)
+    for a in Application.objects.all():
+        ApplicationGenerator(a).generate()
+    management.call_command('make_and_run_migrations')
+    restart(True)
+    return HttpResponse(json.dumps('Running Generations & Migrations...'))
+
+def migrate(request):
+    management.call_command('make_and_run_migrations')
+    return HttpResponse(json.dumps('Making and Running Migrations...'))
